Removing subscriber_descriptions from the status catalog
Previously the show_request command shared a status catalog entry
with the command it was monitoring.  To avoid the show_request
command from overwriting the command description of the monirored
command it sotred its description in a seperate field.  The audit
id was used to tell if create_description was being called by the
show_request command; where the audit id of the requrest didn't
match that of the status catalog entry.

The commit removes that spagetti!

Change-Id: Id96453b8fdbde49a0cffddf1863998cd003ecefc
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>