Move loggable argument description to the status catalog
There are three places where the arguments to a command are processed
and then logged; the status catalog, the log, and the database.  The
first of these, status catalog and log, were half combined in _audit.

This commit moves the creation of a description (as seen by show
active commands) and abridged arguments (as used in the logger)
into the status catalogs create_description.

Change-Id: I191d8e340015987015fa08db7b1ef9fe472c88a4
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>