Seperate input.xml parsing from semantic actions on RestServer
This change seperates the current usage of these data in input.xml by
RestServer into a subclass of CommandRegistry and CommandEntry.  This
leaves CommandRegistry and CommandEntry containing only generic parsing
code.

The classes ResourcesCommandRegistry (which populates RestServer) and
ResourcesCommandEntry (which locates instances of BrokerCommand) now
contain all of the parsing logic.

Change-Id: I4caacfef3bbeb32b8198ae94a8ea2631dc6e461f
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>