Drop the --all option of search commands
The --all option should be handled by the matching show_* command. The
exception is search_hardware, where we don't have a single show_hardware
currently, so leave "search hardware --all" in place for now. Tweak the
definition of search_network a bit so the --network_environment option
can be used alone.

Change-Id: I8db94c6dcea4124e176ad7df086e37926dc8aa77
Reviewed-by: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>