Prevent "aq compile" flushing out reverted state
Previously, the template code tried to restore the timestamp of
plenaries that were rolled back, if compiling them failed. The idea was
that since the profiles did not change, we can avoid re-compiling them.
However, that is not true if we operate on many object templates in a
batch, and not all of them fail to compile. The ones that do compile
will get cached under $quattordir/build/$domain, and the next domain
compile will push those out, because the timestamp is newer than what's
under $quattordir/web/htdocs/profiles.

Not restoring the timestamp will mean the domain compile will re-compile
everything, so we can avoid pushing state out that does not match the
database.

Change-Id: If3fef071f6b29aa194ba3114a0f3a04bb9bd38e1
Addresses-Issue: Jira/AQUILON-1828
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>