Move hardware feature inclusions to the host plenary
In the past we had lock contention issues when we had to update the host
plenaries if the bindings changed. We now have more granular locking, so
that should not be an issue anymore.

This change finally fixes binding features to NIC models; the previous
implementation just did not have enough information to do it properly.

Features which are bound multiple times (e.g., both to the archetype and
to the personality, or both to the name of the NIC and the model of the
NIC) are no longer included multiple times.

Change-Id: Ib1b3be91a0d13ba482b217036090f14b5eb8fff5
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>