Work around Java timestamp granularity issue
The documentation of File.lastModified() says the returned value is in
milliseconds, but the actual implementation in Java 7 (and before)
provides 1 second granularity only. The PAN compiler uses
File.lastModified() to decide if an object needs to be re-compiled. So
if an object is compiled within a second, and a source template is
modified inside the same second, then panc will not notice that the
object needs to be recompiled.

Work around the issue by sleeping a bit if the compilation finished in
the same second when it started. This may also solve some sporadic
compilation-related testsuite failures we haven't been able to track
down earlier.

Change-Id: I13c9b3e28a8819e283bd090a6770fadfecbe8c77
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>