Locking fixes
SELECT ... FOR UPDARE in Oracle won't return rows that were created
after the query started. So when using that construct to protect from
concurrent allocations within the same table, we need to execute two
queries - first one for taking the lock, the second for getting a
consistent result.

This problem does not exist if we can use a row in a different table as
a lock.

Addresses-Issue: Jira/AQUILON-1957
Change-Id: Ice1392592d1af5d61519e7c4a4c30bef86a478b3
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>