support cross environment alias
This allows cross-dns-environment alias to be created. Users can
optionally specify the dns enviornment of the alias target if they
want to create alias which points to target in different dns
environment. (option name: --target_environment)

In order not to confuse users, the raw dns template is also updated
so that if the alias and the target is in different dns environment,
the output will include the alias/target dns environment.

Change-Id: I89f0ffc25a92cd70b3d423a542b3e677224c0a3d
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>