aq.py: Improve status reporting
Use a Condition() to ensure the show_request is sent after the real
command, instead of busy wait/retry. Note that this change does not
guarantee that the server will actually receive and process the real
command before show_request, but it increases the chance that we will
not have to retry the show_request call.

Add linear backoff to the waiting time if we do have to retry.

Change-Id: Iad1c393d72b25808492ff253479d8317a8e0e2fe
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>