Re-work protobuf formatter chain
Protobuf output is a bit tricky: "show" and "search" commands may return
multiple results, so the data type of the result message is always a
list. Previously, protobuf formatter methods dealing with a single
object only were responsible of formatting that object as a list of
length 1. However, that breaks the redirect_proto() mechanism if we want
to delegate the formatting of a singular field. The result was that most
of the protobuf formatting code ended up in
lib/aquilon/formats/formatters.py as helper functions, instead of the
formatter classes where they really belong.

This change introduces a new method, fill_proto(), which is used to
populate a single message only, never a list. The implicit conversion of
single objects to lists is now done in ObjectFormatter when needed.
This allows redirect_proto() to work as expected, and makes it possible
to move most of the protobuf helper functions out of the ObjectFormatter
class.

Change-Id: I402ab8cc663d77a55c1081773f160c305cdb8eb4
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>