Fix binding features to an archetype
Trying to bind a feature that has parameters to a whole archetype
results in obscure errors (e.g. claiming that an empty personality name
is not unique). This change fixes the symptoms which the following
commits would expose, but it does not address cleaning up personalities
if such a feature is then unbound.

Change-Id: Id1b21320ac80dd3e04ae1056c1e2dd7709ef8099
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>