Fix for "show_host --grns --format proto", part 1
The protobuf output generated by "show_host --grns" is not consistent
with the output without the --grns flag: the .hostname field should
contain the short name only, not the FQDN. Fixing it needs multiple
steps - this change causes the .fqdn field to be filled in, so
applications can be updated to use it instead of .hostname. When all
apps are updated, we can fix .hostname to be the short name only.

Change-Id: I7803af7e23eaf81c8d231d3f62340f961f2d0431