Logging improvements
Adapting run_command() to be used in by the consistency checker caused
output being duplicated when run by the broker. Fix that by introducing
a separate stream_level option, which indicates if the caller wants the
output to be streamed to the logger while the command is still running.
As an optimization, fall back to p.communicate() if no streaming is
needed.

When streaming the output to the logger, use the Twisted logging context
from the caller to make sure all lines are prefixed with the request
identifier. Drop the duplicate logging of the output, as this is now
redundant.

Change-Id: I4d23c6ec215c1d15f3a2b73d7818f743f0d43379