Rework service address modelling
Service addresses no longer appear in the address_assignment table.
There's a distinct table for binding service addresses to interfaces
instead, but that is used only if the service address is bound to a
host. If the service address is bound to a cluster, then we expect the
cluster configuration to figure out on which interfaces should the
service address appear.

This change relaxes cluster deco rules, as it is now possible to delete
all nodes of a cluster even if the cluster still holds service
addresses.

Change-Id: I8dc2de71ebb1e334a7bdd963b4e91b644d8dd159
Addresses-Issue: Jira/AQUILON-2146
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>