Refactor DSDB synchronization of service addresses
This is in preparation of the upcoming model changes. Service addresses
are not attached to the primary host record in DSDB, so there's no
reason for using the snapshot_hw()/update_host() machinery, the DSDB
entries can be created/deleted directly. The only exception is the
primary name of Zebra hosts.

After this change, host-bound and cluster-bound service addresses are
treated the same, meaning no interface name is sent to DSDB. The primary
name of Zebra hosts will use 'vip' as the interface name.

Change-Id: Iad3f9a73f86bb9cd63d4843828ee4b14e1da6bf6
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>