Plenaries: Clean up handling of non-compileable archetypes
Do not generate plenaries for non-compileable archetypes by default.
This change fixes a discrepancy where add_host would not generate a data
plenary for a new host, but update_machine would.

Cluster client plenaries will still be generated even if the cluster
itself is not compileable, since members may still use the information.

It can be argued that non-object plenaries should be generated even for
non-compileable archetypes; we may want to revisit that in the future.

Change-Id: I7b9045d61685927887f60dfe14d94c812b13b535
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>