Implement better control of IncompleteError
Add the allow_incomplete flag for commands to express if they care about
IncompleteError being thrown by the plenaries or not. Only commands
calling the PAN compiler do actually care, because the plenaries need to
exist at that point; the common case is to just remind the user that an
"aq reconfigure" will be needed.

Change-Id: I1c37f26bbbf5a1d5bcccb360a3f5123ad67d2a25
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>