Fix cluster enforcement in "aq make"/"aq reconfigure"
"aq make"/"aq reconfigure" did not enforce the cluster_required setting
of the target personality, because the IncompleteError thrown by the
templates was silently converted to a warning.

"aq reconfigure --list" handles this case correctly by returning an
error.

The fix is a bit hacky. The upcoming rework of IncompleteError handling
will make this hack unnecessary.

Change-Id: Id82e98b4919a0786b913255b2070f8fb2d4b5af3
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>