Split parameter definitions per template
This allows better grouping of parameter definitions per the target
template, and will eventually allow better schema definitions.

Currently, the management of the per-template parameter definition
holders is hidden. In the future, we may want new templates to be
registered explicitly, because that would reduce the likelihood of
typing errors.

Change-Id: Id11fa0f7c85fc095452903ca99d0a2a36ab25b89
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>