Add network compartment on its own
The following commit adds network compartment to the broker as
its own entity.  The next commit will add a realation to this
and networks.

For networks compartmentalization is the establishment of boundaries
between networks having different security constraints.  That is to
say that a firewall exists between two networks in different network
compartments.

Change-Id: I9175b761c7465868b1d2c0f261b86d2c90c86320
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>