Fix show_network query options
When show_network is run without --hosts the routers and dynamic stubs
are always printed, yet the data was not automaitcally loaded during
the query by default.  In addition the proto buf output should always
emmit the routers to follow the same pattern as the raw formatter.

The result of this change is that the routers and dynamic_stubs are
always loaded, which makes the query slower.  However, the data are
small (compared to the assignments, then main reason for the optimization).

Change-Id: I5dd60e645783c51c72cff3b86e263d12a8e8ddb7
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>