Feature parameter updates
Parameters are bound to the feature/personality pair, not to a specific
model, so don't pass "--model" to
add_parameter/del_parameter/update_parameter. Add the --type option
instead, to allow disambiguating between features with the same name but
different types.

Also drop the --interface_name parameter, as it would not work well with
binding interface parameters based on the model of the interface.
Per-interface parameters can still be implemented using JSON objects
indexed with the interface name, if needed.

Change-Id: I2e6064a0027125a384c758dd5dc0fabf59b4255a
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>