Get rid of requires_azcheck
requires_azcheck was set to False by the same commands which set
requires_transaction to False. So simplify the rules - if the command
needs a database session, then it will go through authorization too.

Also drop the session option of invoke_render(). That was an artiface
from before factoring out invoke_render() - contrary to the render()
method, calls to invoke_render() should not be nested, so nothing should
pass an existing session to it.

Change-Id: I00a58a004bfc58261006c2da06c36a190c0ed333
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>