Make the sleep after compilations unconditional
We need to ensure the timestamp of the last output file generated by the
compiler will be different then the first command which will execute
after it. Since we don't know when did the compiler write the last
output file, trying to be smart about the delay does not work in
practice.

Change-Id: Ic7c65921ce4d2c50b500785d96c3d24f22bcbb2f
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>