Enforce archetype parameters to start with the template name
The parameter handling code discards the first component of the path
once the matching definition was found. This allows creating multiple
parameter definitions which "alias" each other, e.g.:

  add parameter definition --template t --path foo/var
  add parameter definition --template t --path bar/var

Both of the above definitions would cause '"var" = ...' to be added to
template t, which would generate a conflict.

Enforcing the path to always start with the template name avoids this
situation, and leaves the meaning of --path compatible with the
add_parameter command.

Change-Id: I5f57ab935cfb0260c98997a7af9e91e237db2898
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>