Prefer the boot interface for default gateway
We need a single default gateway to be selected for basic OS
configuration. If there are multiple interfaces configured to provide
the default route, then prefer the boot interface, and defer configuring
multipath routing to the templates.

This fixes a potential instability in the tests if the DB returns the
interfaces/addresses in a different order than what we expect.

Change-Id: I71648e073f88adb51b6f05a4fd77a3f0defe9bc2
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>