Relationship cleanups
Clean up the use of dns_fqdn_mapper. Use the mapper itself for accessing
the columns, instead of addressing the objects behind it. Use a
pre-defined join() instead of accessing tables directly.

Use remote()/foreign() annotations when that makes the code easier to
read.

Clean up which side is foreign and which one is remote - this is
somewhat arbitrary as far as the SQL conditions are concerned, but
SQLAlchemy uses this information to determine which should be the "many"
side, so we don't have to use "uselist=..." annotations.

Change-Id: I8fcfaadeb59b34762c9b09c8f6d8a141a5d2e254
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>