Refresh network plenaries more
add_interface_address/del_interface_address may associate/disassociate
an IP address with a router, so let's refresh the network plenaries. We
could make that conditional on the IP in fact being a router address,
but the data in the network plenary needs to be loaded for generating
the host plenary too, so it's unclear if the extra lookup would be a win
or not.

Make update_router_address also refresh the network plenary, to be in
line with the other update commands.

Change-Id: I2577dfe313037f85d70756c13fbd4794edf583bb
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>