Flatten archetype parameters
After this change, the JSON blob stored for an archetype parameter
template becomes semantically equivalent with the Pan structure template
being generated, so the conversion is purely syntactical. So, if
functionality to read JSON objects is added to panc, then we'll be able
to dump the value stored in the DB as-is to the filesystem, and could
avoid doing any kind of conversion.

Change-Id: I4bbcc47128403b91d843e107f6f4de9a3906fdf7
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>