Fix parameter cleanup when a feature is unbound
There may be multiple links between the feature and the personality
(e.g. a hardware feature which is bound to multiple models), so the
parameters need to be kept until the last link is removed.

Do the rebuild check only if a value was actually set for such a
parameter.

Change-Id: Id2a04d3d6c330354e1096a9f28103d293fd86972
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>