Make sure the plenaries for the next stage get created on disk
For staged personalities, the Personality.active_stage property creates
the 'next' stage on the fly, if needed. There are a couple scenarios
(required personalities, static routes, hardware feature bindings) when
the relation we want to work with is bound to the PersonalityStage
object, but the information is not exported to the personality plenaries
(it may be exported to the host plenaries, or it may not be exported at
all). In those cases, we want to flush the personality plenaries if the
stage was newly created, so moving a host to the 'next' stage won't fail.

Change-Id: I472af55e567dd8920d61a44a16237d1ae513e1c5
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>