Add support for an archetype preface template
This change introduces support for a "preface" template which provides a
mechanism to modify the LOADPATH before any other templates are processed.

In order to change the version of the template library being used by an
archetype, the LOADPATH must be modified before any other templates are
included, in particular pan/units and pan/functions, which are included before
archetype/base.

This also sets an inital value for the global ARCHETYPE variable.

Change-Id: Ic67bbabb0addcda0e8d83c62860f352aaf0e817d