Make port group display more verbose
Show if a network is bound to a port group. If an interface is bound to
a port group, then display the network too, to make it easier to spot if
the interface has an IP address assigned from the wrong network (which
is a warning condition currently, not a hard error).

Change-Id: I7adb7662e8b96c249308389462d3a7db53cb03ce
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>