Avoid duplications in EON ID maps
Broker-exported GRN maps work only for personalities which do not
override config.tpl. So far, the host plenary included a mapping which
mixed host-level and personality-level maps, to allow working with
legacy personalities. Since such personalities are now gone, we can
remove this compatibility hack, which avoids duplication of entries in
the profile.

Change-Id: If0043f1cb240270605b0db4c09fb183cb9fcbe54
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>