Drop content from PlenaryCollection subclasses
PlenaryCollection is used as a dual role: as an arbitrary collection of
random other plenaries, and as the coordinator for writing/updating
multiple templates at once. To be able to improve the second role
further, we need to get rid of the first. The first step toward that
goal is removing any extra internal state added by the subclasses.
Thankfully, that was already mostly unused, only "aq cat" of
resourcegroup objects needed some adjustments.

Change-Id: I15406d1f01536c7922c2a2ce88c569dd1edb594e
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>