Move the instantiation of PlenaryCollection to BrokerCommand
Eventually, we'd like to track plenary changes in git, which will
require passing more metadata about the request to PlenaryCollection. By
moving the instantiation of PlenaryCollection to BrokerCommand, we can
avoid having to touch each and every command later.

Some users of this new feature needed to be converted manually to avoid
issues due to **-style parameter passing, the rest will be converted
automatically.

Change-Id: I657ec7d502eeb00142830e22ceaf76e526a00210
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>