Network method name changes
The ipaddress module is now standard in Python 3, and it has some
differences compared to the ipaddr module. Modify some properties of our
Network class to be more similar to ipaddress.

Eventually, it would be nice to rename the Network.ip column to
Network.network_address, and Network.cidr to Network.prefixlen.

Change-Id: I0d2ff4d8e33f92dd9a9d7a896823974cc3816b71
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>