Temporarily remove change-validation for make_cluster.
This temporarily removes the validate_prod_cluster call in the
make_cluster command -- with the change-management code as it
stands, calls that expect to require justification in the future
require the parameter to be present, even though it is ignored.
This broke virtualisation's tooling in the last turnover.

Change-Id: I4b27163a4e804faa92b5bd6778d7b49ad7a1fb93
Addresses-Issue: Jira/AQUILON-3821
Reviewed-By: Dave Reeve <Dave.Reeve@morganstanley.com>