Process the request ID early
Previously, if one of the arguments could not be interpreted, then the
client got some ugly looking messages about the request not being found.
Processing the request ID early allows recording the request in the
status catalog before the rest of the arguments are processed, so errors
will be handled as expected.

BrokerCommand.add_logger was renamed to pre_render, because it does not
have much to do with logging setup anymore.

Change-Id: I63561ce87567b93b2cb7baee9d112dd0e4166c0b
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>