Protocol and parameter encoding fixes
Add more charcater encoding enforcement to the request processing: the
protocol should handle UTF-8 (and no other encoding), even if we're not
ready to advertise that yet, and we still restrict command options to
the ASCII character set later.

Also fix a few places where encode()/decode() was used in the wrong
direction. It is not really a big deal with Python 2 because both the
argument and the result is str and all we want is to check if the
encoding is valid, but Python 3 does care about str vs. bytes.

Change-Id: Ib6a09373316b82f0bd1ba698a6429f249de3c659
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>