Support configurable mode on hostlinks.
add_hostlink --mode

If Aquilon creates the target of the hostlink (ie. it is on local
storage) then it will create it with this mode.

The templates copy it from the /system/resource/hostlink entries into
the dirperm entry it creates. This change simply stores the mode in the
database and outputs it to the hostlink plenary.

Change-Id: I7582e1a690e227ea877249992d7d970bbaadb713
Addresses-Issue: Jira/AQUILON-3973
Reviewed-By: Fred Barnes <Fred.Barnes@morganstanley.com>