Support setting a uri on buildings.
This field is intended to be used as a reference to some external
system, where the relationship with the building needs to be modelled.

It should be a string in the format system://id, allowing to distinguish
between external systems.

It can be validated by the external tool if location_uri_validator is
provided in the config.

Change-Id: I36eeb3bab48c1cfd1be1a3250a4c71ee5baa1e37
Addresses-Issue: Jira/AQUILON-4011
Reviewed-By: Fred Barnes <Fred.Barnes@morganstanley.com>