Disallow duplicates in --list arguments
This updates the broker's force_list() routine to check and disallow
duplicates, as given in 'list' type arguments.  The reason is that when
the audit log is created, these end up as individual entries in the
XTN_DETAIL table, with duplicate entries generating a primary-key
constraint violation error.  Moreover, this probably indicates a mistake
in the originally provided list.

Change-Id: I8086eee60756c31bef85b32e7f89c725d5641178
Addresses-Issue: Jira/AQUILON-4282
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>