Serialise concurrent add_machine_prefix with same prefix
This tries to ensure that concurrent invocations of add_machine_prefix
with the same prefix, for the first time (where there are no rows to
lock), doesn't result in an IntegrityError exception.  This is handled
by locking the first row of the table (ordered by label) in the case
where no prefix already exists.

Change-Id: I7f0ca05e8fa941e1def4742db64971cf11774f56
Addresses-Issue: Jira/AQUILON-4279
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>