notifications: add missing DNS update notifications when creating/deleting a zebra host
When creating a zebra host, the reverse DNS records (PTR) of the IP
addresses of the interfaces are modified to be asymmetric and to all
point to the primary hostname of the host. This patch adds the
notification to inform of that update. It also adds a notification
when deleting such host, and the reverse DNS records fall back to
be symmetric.

Change-Id: I3f03e89dffb22dad054bd8626a54c0efcb6e7c8e
Addresses-Issue: Jira/AQUILON-4545
Reviewed-By: Gintare Urbone <gintare.urbone@morganstanley.com>