fix: 'dump dns' proto format should only return one DNSRecord object per FQDN
As the FQDN returned when querying the database are not necessarily
sorted, the 'dump dns' proto format was returning a DNSRecord
objects for each record, and thus multiple DNSRecord objects for a
single FQDN, containing each a single DNSRecordData object.
This fixes that behavior to return a single DNSRecord object per FQDN,
containing a DNSRecordData object for each record.

Change-Id: I0bf0d155d58bfd0d4e4fba0b5afd5c86abe147ba
Addresses-Issue: Jira/AQUILON-4631
Reviewed-By: Gintare Urbone <gintare.urbone@morganstanley.com>