Improve the help text for the publish command
Add a note to make it clear that the publish command accepts a branch
name without the $USER/ prefix that is required for some other
sandbox-related commands.

Clarify the wording used to describe the --rebase option to use git
terminology that is now in fairly common usage.

Change-Id: I6012c62ca3f8132df524d6e30b34ae2f65ba361d
Addresses-Issue: Jira/AQUILON-941
Reviewed-By: Gintare Urbone <gintare.urbone@morganstanley.com>