Fix: show dynamic range returns "NotFound" for IPs outside of a dynamic range
The show dynamic range command was returning a range from and to
the queried IP, with a size of 1, even though the IP was not in a
dynamic range. This patch fixes that behavior by raising a
NotFoundException if the IP is not in a dynamic range.

Change-Id: Ib9e3c3286bb07246ad5554f510d2bd0a4dd49b84
Addresses-Issue: Jira/AQUILON-4775
Reviewed-By: Gintare Urbone <Gintare.Urbone@morganstanley.com>
Reviewed-By: Fred Barnes <Fred.Barnes@morganstanley.com>