aq.py: add check in default options for 'host' value
Other options such as 'aqservice' and 'port' are set by checking,
among others, the 'defaultOpts' dict (filled with the configuration
file), but for 'host' we were only checking the 'globalOpts' dict
(corresponding the global options in the command line). This fixes that.
This also removes the mention to 'get_username()' to set 'aqservice'
as it is already done in the if-else block setting the 'default_aqservice'.
Fixes #96 (GitHub issue https://github.com/quattor/aquilon/issues/96)

Change-Id: Ib70cde2d494f2afe5eaa6546d84a8e49f4107a75