Add an 'extend' action for the command line parameters
The 'multiple' type does not allow to manage typed parameters,
this thus introduces the 'extend' action that allows, while
put together with any of the basic parameter types, to repeat
the parameter multiple times with different values, or separate
the values using the comma (,) character. The data will be
converted to JSON for the transfer from the client to the broker,
then converted back on the other side.

Change-Id: Ia3ab08f33fe0a373fe17bb45d89144a20985f226
Reviewed-by: Gintare Urbone <Gintare.Urbone@morganstanley.com>
Tested-by: Aquilon Template Build testing and verification