entitlements: add entitlements tables and related commands
This adds the database tables to store the entitlements to
user and GRNs on hosts, clusters, personalities, archetypes
or GRNs. This also adds the related commands to add, del,
show and search entitlements. The show command allows to
show any entitlements that apply, directly or by inheritance,
to the host given as parameter.

Change-Id: Ic460378d50171d409b58b1bb826b6598817f3c9d
Addresses-Issue: Jira/AQUILON-5210
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>