entitlements: write the entitlements in plenaries
This adds a /system/entitlements/<type> parameter in the
plenaries of the objects targetted by the entitlements,
where <type> is the entitlements type.
This parameter will contain a dict with two keys: 'eon_id',
and 'user', each containing a list of struct of the
elements that need to be entitled for the given type.
The struct will contain a 'value' (name of the user or
numeric value of the eon id) and a 'type' (only applicable
for a user to specify the user type).

Change-Id: I2081948320d2e0b1275a9d3e9b2bbb98835672e9
Addresses-Issue: Jira/AQUILON-5210
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>