worker: templates: add plenaries for ParameterizedGrn
Add plenary class as well as "cat --grn" and "cat --eon_id"
commands, to be used with "--host_environment" as well as a
location parameter, to manage plenaries at the Grn level.
This also adds a PlenaryHostGrn object to load all the
parameterized GRN plenaries that apply to a given host,
and a "--host_grn" option to use with "cat --hostname"
in order to see this plenary.

Addresses-Issue: Jira/AQUILON-5331
Change-Id: I7d1a566bb13aa07abe823c9ca94c078f5940242e
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>