worker: templates: add plenaries for ParameterizedArchetype
Add plenary class as well as "cat --archetype" command, to be
used with "--host_environment" as well as a location parameter,
to manage plenaries at the Archetype level.
This also adds a PlenaryHostArchetype object to load all the
parameterized archetype plenaries that apply to a given host,
and a "--host_archetype" option to use with "cat --hostname"
in order to see this plenary.

Note that the archetype option of the cat command was already
available in the cat command, and keeps the behavior of
being the personality's archetype when used together with the
--personality option of the command.

Change-Id: I0bf5e38be35eb2fa372fec8085da6e114473d566
Addresses-Issue: Jira/AQUILON-5332
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>