refresh_user: use a source script instead of a source list
The list of users was provided through a source list
until now, which meant that the user information could
not come from anything else than a file, hence requiring
us to have that file ready when running the command.
This updates that behaviour by using a source script
instead; Anyone wanting to still use a file can basically
'cat' it in a script and feed that script to Aquilon.

The expected input format also changed a bit, as we now
expect a well-formatted CSV, comma-separated, with a
header line. The header line will be used to parse the
CSV. If a 'type' column is provided, it will be used as
the user type. If no 'type' column is provided, all user
types will default to the one defined in the configuration.

Change-Id: Ifeb34326dd6f50612eb93c18b4040a8fbabd8423
Addresses-Issue: Jira/AQUILON-5273
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Cesar A. Sanchez <Cesar.A.Sanchez@morganstanley.com>