templates: base: only try to release key if a key has been found
In cases where the get_key() method to find the template key to
lock was raising an exception, the key parameter is thus staying
set to None. However, in the finally block, only the need for
being locked was checked before trying to unlock as key. This is
fixed by also checking that the key we are trying to unlock exists.

Change-Id: I00f8317089c0d41d8504622bce681dfdeba29624
Reviewed-by: Tomasz Kotarba <Tomasz.Kotarba@morganstanley.com>
Tested-by: Aquilon Template Build testing and verification