Add parameterized objects for GRN, Archetype and Personality
Parameterized objects are a wrapper around a main object that
allows to specify other objects that specify the parameterization
of the main one. This includes a ParameterizedGrn which is
parameterized with an host environment and a location, a
ParameterizedArchetype with the same parameters, and a
ParameterizedPersonality per location.

Change-Id: I17a6efa8c126cdf84dba5ed79b64aca786403d3b
Addresses-Issue: Jira/AQUILON-5210
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>