worker: templates: add plenaries for ParameterizedPersonality
Add plenary class as well as "cat --personality" command, to
be used with a location parameter, to manage plenaries at the
parameterized personality level.
This also adds a PlenaryHostPersonality object to load all the
parameterized personality plenaries that apply to a given host,
and a "--host_personality" option to use with "cat --hostname"
in order to see this plenary.

Note that the "cat --personality" command already existed
previously and keeps its previous behavior except when used
together with a location parameter.

Addresses-Issue: Jira/AQUILON-5210
Change-Id: Ifa71c2130b6d1417f1e5fe78279d4422b8e02371
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>