optparser: improve requirements and conflicts handling
Previously, requirements and conflicts were handled
separately and with a different logic, this merges both logics
together and allows for conflicts inheritance the same way
requirements inheritance was allowed previously.
This also introduces the possibility of joined requirements
or conflicts through the use of the '+' symbol: if requirements
are set to "a+b c+d e", it would mean that we require either
"a" _and_ "b", or "c" _and_ "d", or "e". Same logic applies
for the conflicts.

Addresses-Issue: Jira/AQUILON-5720
Change-Id: I72d6b52c9dedce26bcfe53d0d830bd9e93960483
Reviewed-by: Cesar A. Sanchez <Cesar.A.Sanchez@morganstanley.com>
Tested-by: Aquilon Template Build testing and verification