Fix misleading verb in show_permission output
This ensures that options for which users lack permission to execute are
correctly marked with 'except with: --option-name' instead of the
erroneous 'expect with: --option-name'.
It also fixes the bug in tests for show_permission so that the tests no
longer require erroneous "expect with" instead of the correct "except
with".

Addresses-Issue: Jira/AQUILON-5432
Change-Id: Iaecd68ae43aa57011c9052e2818a60f878841666
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>
Reviewed-by: Raphael Beamonte <Raphael.Beamonte@morganstanley.com>