tests/codestyle: fix the --current option to blame properly
The --current option allows to consider the files in the
selected commit(s) but blame them against the current state
of the file. This was not behaving properly as the git blame
was considering only data up to the last selected commit,
and ignoring the currently uncommitted code lines.
This fixes that.

Change-Id: I04b85ca246f790d5071f25fcfde0a27eda263d52
Addresses-Issue: Jira/AQUILON-5722
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Alexander Jackson <Alexander.Jackson@morganstanley.com>