search_dns: allow for --(exclude|all)_(network|dns)_environment(s)
This provides a way to exclude a network environment (and consequently
a dns environment as they are strongly linked) from the results of a
search_dns query. This also allows to provide multiple network
environments (or dns environments) to include or exclude in the results.
The --all_(network|dns)_environments flags have also been added to get
results for all network/dns environments instead of filtering in/out.

Change-Id: I15f6e42906e97cf14c1da97af2fef55bd2076e26
Addresses-Issue: Jira/AQUILON-5848
Tested-by: Aquilon Template Build testing and verification
Reviewed-by: Fred Barnes <Fred.Barnes@morganstanley.com>